+++ CHANGELOG.md
@@ -26,8 +26,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 1. Update AsyncHttpClient to change from Vertx to Reactor HttpClient
-2. Upgrade reactor-bom to version 2024.0.9 that fetches reactor-netty-core version 1.2.9
-3. Upgrade netty to version 4.2.4.Final
+2. Change the concurrency of AsyncHttpClient and TemporaryIndex to 500, matching the underlying HTTP client's connection pool
+3. Upgrade reactor-bom to version 2024.0.9 that fetches reactor-netty-core version 1.2.9
+4. Upgrade netty to version 4.2.4.Final
 
 ---
 ## Version 4.3.12, 8/11/2025